requirements.txt
paho-mqtt
requests
pytest
pytest-html